update details.php and other files

+++ cars/details.php
@@ -9,33 +9,53 @@
     $result = mysqli_query($connect, $sql);
     $row = mysqli_fetch_assoc($result);
 
-    if(!isset($_SESSION['adm'])){
+
+    if(!isset($_SESSION['adm']) && !isset($_SESSION['user'])){
        $body .= "
-        <div class='card m-1' style='width: 48rem;'>
+        <div class='card cardBody m-4'>
             <img src='./img/{$row['picture']}' class='card-img-top' alt='{$row['brand']}'>
             <div class='card-body'>
                 <h5 class='card-title'>{$row['brand']} {$row['model']}</h5>
-                <p class='card-text'>Price per day: {$row['pricePerDay']}  &#8364;</p>
-                <p class='card-text'>Type of fuel: {$row['fuelType']}</p>
-                <p class='card-text'>Number of seats: {$row['seatNum']}</p>
+                <p class='card-text'>{$row['pricePerDay']}  &#8364; per day</p>
+                <p class='card-text'><i class='bi bi-fuel-pump'></i> {$row['fuelType']}</p>
+                <p class='card-text'><i class='bi bi-person'></i> {$row['seatNum']} seats</p>
             </div>
-    <   /div>"; 
-    } else {
+        </div>"; 
+    } elseif(isset($_SESSION['user'])){
         $body .= "
-        <div class='card m-1' style='width: 48rem;'>
+        <div class='card cardBody m-4'>
             <img src='./img/{$row['picture']}' class='card-img-top' alt='{$row['brand']}'>
             <div class='card-body'>
                 <h5 class='card-title'>{$row['brand']} {$row['model']}</h5>
-                <p class='card-text'>Price per day: {$row['pricePerDay']}  &#8364;</p>
-                <p class='card-text'>Type of fuel: {$row['fuelType']}</p>
-                <p class='card-text'>Number of seats: {$row['seatNum']}</p>
+                <p class='card-text'>{$row['pricePerDay']}  &#8364; per day</p>
+                <p class='card-text'><i class='bi bi-fuel-pump'></i> {$row['fuelType']}</p>
+                <p class='card-text'><i class='bi bi-person'></i> {$row['seatNum']} seats</p>
+                <div class='btnBox'>
+
+                <a href='./booking.php?id={$row['id']}' class='btn btn-success'>Book this car</a>
+            </div>
+            </div>
+        </div>"; 
+    }else {
+        $body .= "
+        <div class='card m-4'>
+            <img src='./img/{$row['picture']}' class='card-img-top' alt='{$row['brand']}'>
+            <div class='card-body'>
+                <h5 class='card-title'>{$row['brand']} {$row['model']}</h5>
+                <p class='card-text'>{$row['pricePerDay']}  &#8364; per day</p>
+                <p class='card-text'><i class='bi bi-fuel-pump'></i> {$row['fuelType']}</p>
+                <p class='card-text'><i class='bi bi-person'></i> {$row['seatNum']} seats</p>
             </div>
             <div class='btnBox'>
+                <a href='./booking.php?id={$row['id']}' class='btn btn-secondary'>Book this car</a>
+                
                 <a href='./update.php?id={$row['id']}' class='btn btn-warning'>Update</a>
                 <a href='./delete.php?id={$row['id']}' class='btn btn-danger'>Delete</a>
             </div>
     <   /div>"; 
     } 
+
+
     
 ?>
 <!DOCTYPE html>
@@ -52,7 +72,31 @@
     <?php require "./navbar.php"; ?>
 
    <div class="container">
-   <?= $body; ?>
+        <?= $body; ?>
+        <div class="content">
+            <h3>Great choice!</h3>
+            <div class="row row-cols-2">
+                <p class="par"><i class="bi bi-check-lg"></i> Customer rating: 7,7 / 10</p>
+                <p class="par"><i class="bi bi-check-lg"></i> Most popular fuel policy</p>
+                <p class="par"><i class="bi bi-check-lg"></i> Short queues</p>
+                <p class="par"><i class="bi bi-check-lg"></i> Easy to find counter</p>
+                <p class="par"><i class="bi bi-check-lg"></i> Helpful counter staff</p>
+                <p class="par"><i class="bi bi-check-lg"></i> Free Cancellation</p>
+            </div>
+            <hr>
+            <h3>Included in the price</h3>
+            <div class="row row-cols-2">
+                <p class="par"><i class="bi bi-check-lg"></i> Free cancellation up to 48 hours before pick-up</p>
+                <p class="par"><i class="bi bi-check-lg"></i> Collision Damage Waiver with €2,100 excess</p>
+                <p class="par"><i class="bi bi-check-lg"></i> Theft Protection with €2,100 excess</p>
+                <p class="par"><i class="bi bi-check-lg"></i> Unlimited mileage</p>   
+            </div>
+            <hr>
+            <h3>Extras</h3>
+            <div class="row row-cols-2">
+                <p class="par"><i class="bi bi-check-lg"></i> Child seats, additional drivers and more.</p>
+            </div>
+        </div>
    </div>
 
    <?php require "./footer.php"; ?>
